user schema에 imagePath 추가

+++ back/src/routers/educationRouter.js
@@ -57,7 +57,6 @@ educationRouter.put("/educations/:id", async (req, res, next) => {
 
     const education = await Education.findByEducationId({ education_id });
 
-    // !!!
     if (education.user_id !== req.currentUserId) {
       throw new Error("권한이 없습니다.");
     }
@@ -89,10 +88,9 @@ educationRouter.delete("/educations/:id", async (req, res, next) => {
     const education_id = req.params.id;
     const education = await Education.findByEducationId({ education_id });
 
-    // !!!
-    // if (education.user_id !== req.currentUserId) {
-    //   throw new Error("권한이 없습니다.");
-    // }
+    if (education.user_id !== req.currentUserId) {
+      throw new Error("권한이 없습니다.");
+    }
 
     await educationService.deleteEducation({ education_id });
